update: change the dir for parser interface
